EmailMessageBuilderTest was really a unit test, so removed dependency to database.
